+++ app/models/credentials.rb
@@ -2,6 +2,7 @@
 
 require 'bcrypt'
 require 'util/cidr'
+require 'conjur/password'
 
 # TODO: This is needed because having the same line config/application.rb is
 # not working.  I wasn't able to figure out what precisely was going wrong,
@@ -13,13 +14,6 @@ Sequel::Model.db.extension(:pg_array, :pg_inet)
 class Credentials < Sequel::Model
   # Bcrypt work factor, minimum recommended work factor is 12
   BCRYPT_COST = 12
-
-  # special characters according to https://www.owasp.org/index.php/Password_special_characters
-  VALID_PASSWORD_REGEX = %r{^(?=.*?[A-Z].*[A-Z])                             # 2 uppercase letters
-                             (?=.*?[a-z].*[a-z])                             # 2 lowercase letters
-                             (?=.*?[0-9])                                    # 1 digit
-                             (?=.*[ !"#$%&'()*+,-.\/:;<=>?@\[\\\]^_`{|}~]).  # 1 special character
-                             {12,128}$}x                                     # 12-128 characters
 
   plugin :validation_helpers
 
@@ -80,7 +74,15 @@ class Credentials < Sequel::Model
 
     validates_presence [ :api_key ]
 
-    errors.add(:password, ::Errors::Conjur::InsufficientPasswordComplexity.new.to_s) if @plain_password && (@plain_password.index("\n") || @plain_password !~ VALID_PASSWORD_REGEX)
+    # We intentionally don't validate when there is no password
+    # See flow in Account.create
+    return unless @plain_password
+
+    unless Conjur::Password.valid?(@plain_password)
+      errors.add(
+        :password, ::Errors::Conjur::InsufficientPasswordComplexity.new.to_s
+      )
+    end
   end
 
   def before_validation